backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
firebase-admin>=6.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0

# Testing dependencies
pytest>=7.3.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
